backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pillow==10.1.0
google-generativeai==0.3.1
diffusers>=0.25.0
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1
opencv-python-headless
scikit-learn==1.3.2
numpy==1.24.3
backoff==2.2.1
langdetect==1.0.9
email-validator==2.1.0
python-dotenv==1.0.0
ip-adapter
aiosqlite 